Arguments have a nullvalue node if they are not initialized
